requirements.txt
# Core requirements
python-docx    # For .docx files
pywin32; platform_system=="Windows"    # For .doc files (Windows only)
pandas        # For Excel files
openpyxl      # For .xlsx files
xlrd          # For .xls files
PyMuPDF       # For PDF files (primary)
pdfplumber    # For PDF files (backup)
openai==0.28.0  # For OpenAI API (specific version)